+++ seleniumpm/__init__.py
@@ -9,7 +9,7 @@ Selenium PageModel library
 """
 
 __title__ = 'seleniumpm'
-__version__ = '2.6.0'
+__version__ = '2.7.0'
 __build__ = 0x021000
 __author__ = 'Peter Salas'
 __license__ = 'Apache 2.0'