+++ seleniumpm/examples/superwikipedia.py
@@ -2,7 +2,6 @@ from seleniumpm.examples.wikipedia import Wikipedia
 from seleniumpm.examples.widgets.WikipediaPersonal import WikipediaPersonal
 from seleniumpm.webelements.link import Link
 from seleniumpm.locator import Locator
-from selenium.webdriver.common.by import By
 
 
 class SuperWikipedia(Wikipedia):
@@ -15,5 +14,5 @@ class SuperWikipedia(Wikipedia):
 
     def __init__(self, driver, url):
         super(SuperWikipedia, self).__init__(driver, url)
-        self.mainpage_link = Link(driver, Locator(By.XPATH, "//li[@id='n-mainpage-description']/a"))
-        self.personal_widget = WikipediaPersonal(driver, Locator(By.XPATH, "//div[@id='p-personal']"))
+        self.mainpage_link = Link(driver, Locator.by_xpath("//li[@id='n-mainpage-description']/a"))
+        self.personal_widget = WikipediaPersonal(driver, Locator.by_xpath("//div[@id='p-personal']"))